Validatedata

+++ ProjectManagementAPI/Persistance/Repositories/TaskRepository.cs
@@ -3,7 +3,6 @@ using ProjectManagementAPI.Domain.Repositories;
 using ProjectManagementAPI.Models;
 using System.Collections.Generic;
 using System.Threading.Tasks;
-using Task = System.Threading.Tasks.Task;
 
 namespace ProjectManagementAPI.Persistance.Repositories
 {
@@ -11,35 +10,54 @@ namespace ProjectManagementAPI.Persistance.Repositories
     {
         public TaskRepository(AppDbContext context) : base(context) { }
 
-        public async Task CreateTask(Models.Task task)
-        {            
+        public async Task<bool> CreateTask(Models.Task task)
+        {
             _context.Tasks.Add(task);
             await _context.SaveChangesAsync();
+            return true;
         }
 
-        public async Task DeleteTask(long id)
+        public async Task<bool> DeleteTask(long id)
         {
             var task = _context.Tasks.Find(id);
+            if (task == null)
+            {
+                return false;
+            }
             _context.Tasks.Remove(task);
             await _context.SaveChangesAsync();
+            return true;
         }
 
         public async Task<Models.Task> GetTask(long id)
         {
-            var task = await _context.Tasks.Include(t => t.Owner).Include(t => t.Project).FirstOrDefaultAsync(t => t.Id == id);
+            var task = await _context.Tasks
+                .Include(t => t.Owner)
+                .Include(t => t.Project)
+                    .ThenInclude(t => t.Owner)
+                    .FirstOrDefaultAsync(t => t.Id == id);
             return task;
         }
 
         public async Task<IEnumerable<Models.Task>> GetTasks()
         {
-            return await _context.Tasks.Include(t => t.Owner).Include(t => t.Project).ToListAsync();
+            return await _context.Tasks
+                .Include(t => t.Owner)
+                .Include(t => t.Project)
+                    .ThenInclude(t => t.Owner)
+                .ToListAsync();
         }
 
-        public async Task UpdateTask(long id, Models.Task task)
+        public async Task<bool> UpdateTask(long id, Models.Task task)
         {
-            task.Id = id;
+            var foundTask = await _context.Tasks.FindAsync(id);
+            if (foundTask == null)
+            {
+                return false;
+            }
             _context.Entry(task).State = EntityState.Modified;
             await _context.SaveChangesAsync();
+            return true;
         }
     }
 }
